modelTemplateMatching/requirements.txt
flask
pillow
numpy
opencv-python
pytesseract
pymysql
werkzeug
